Group activity 2

+++ src/main/java/io/muzoo/ssc/activityportal/backend/activity/ActivityServiceImpl.java
@@ -19,6 +19,7 @@ import java.util.List;
 public class ActivityServiceImpl implements ActivityService {
     @Autowired
     private ActivityRepository activityRepository;
+
     @Autowired
     private GroupRepository groupRepository;
     @Autowired
@@ -26,10 +27,8 @@ public class ActivityServiceImpl implements ActivityService {
     @Autowired
     private WhoamiService whoamiService;
 
-    @Autowired
-    public ActivityServiceImpl(ActivityRepository activityRepository) {
-        this.activityRepository = activityRepository;
-    }
+    private static final long  deleteAfterCompletionTimePeriod = 60000; // 1 minute
+
 
     public SimpleResponseDTO checkUserAndGroup(long groupId) {
         User u = whoamiService.getCurrentUser();
@@ -46,12 +45,21 @@ public class ActivityServiceImpl implements ActivityService {
         return SimpleResponseDTO.builder().success(true).message("User and group checked").build();
     }
 
-    @Scheduled(fixedRate = 60000) // Updates every minute
-    public void updateActivityStatus() {
+    @Transactional
+    @Scheduled(fixedRate = 600000) // 10 minutes
+    public void updateAndDeleteActivityStatus() {
         LocalDateTime now = LocalDateTime.now();
-        // Get all activities
         Iterable<Activity> activities = activityRepository.findAll();
         for (Activity activity : activities) {
+            if (now.isAfter(activity.getEnd_time())&& now.isAfter(activity.getEnd_time().plusMinutes(1))){
+                // Delete the activity if it is completed
+                Group group = activity.getGroup();
+                for (User user : activity.getUsers()) {
+                    user.getActivities().remove(activity);
+                }
+                group.getActivities().remove(activity);
+                activityRepository.delete(activity);
+            }
             if (now.isBefore(activity.getStart_time())) {
                 activity.setStatus("PENDING");
             } else if (now.isAfter(activity.getStart_time()) && now.isBefore(activity.getEnd_time())) {
@@ -61,6 +69,27 @@ public class ActivityServiceImpl implements ActivityService {
             }
         }
     }
+
+    private SimpleResponseDTO validateActivity(Activity activity, long groupId) {
+        // Check if the activity exists
+        if (activity == null) {
+            return SimpleResponseDTO.builder().success(false).message("Activity not found").build();
+        }
+        // Check if the activity belongs to the group
+        if (activity.getGroup().getId() != groupId) {
+            return SimpleResponseDTO.builder().success(false).message("Activity does not belong to the group").build();
+        }
+        // Check if the activity is ongoing
+        if (activity.getStatus().equals("ONGOING")) {
+            return SimpleResponseDTO.builder().success(false).message("Activity is ongoing").build();
+        }
+        // Check if the activity is completed
+        if (activity.getStatus().equals("COMPLETED")) {
+            return SimpleResponseDTO.builder().success(false).message("Activity is completed").build();
+        }
+        return SimpleResponseDTO.builder().success(true).message("Activity validated").build();
+    }
+
     @Override
     @Transactional
     public SimpleResponseDTO createActivity(Activity activity, long groupId) {
@@ -77,32 +106,28 @@ public class ActivityServiceImpl implements ActivityService {
                 userRepository.save(member);
             });
             // Update the status of the activity
-            updateActivityStatus();
+            updateAndDeleteActivityStatus();
             return SimpleResponseDTO.builder().success(true).message("Activity created").build();
         }).orElse(SimpleResponseDTO.builder().success(false).message("Group not found").build());
     }
 
+    @Transactional
+
     public SimpleResponseDTO editActivityDetails(Activity activityDetail, long groupId, long activityId) {
+        Activity activity = activityRepository.findFirstById(activityId);
+        validateActivity(activity, groupId);
+
         SimpleResponseDTO checkResult = checkUserAndGroup(groupId);
         if (!checkResult.isSuccess()) {
             return checkResult;
         }
-        Activity updateActivity = activityRepository.findFirstById(activityId);
-        // Check if the activity exists
-        if (updateActivity == null) {
-            return SimpleResponseDTO.builder().success(false).message("Activity not found").build();
-        }
-        // Check if the activity belongs to the group
-        if (updateActivity.getGroup().getId() != groupId) {
-            return SimpleResponseDTO.builder().success(false).message("Activity does not belong to the group").build();
-        }
-        updateActivityStatus();
-        updateActivity.setName(activityDetail.getName());
-        updateActivity.setDescription(activityDetail.getDescription());
-        updateActivity.setStart_time(activityDetail.getStart_time());
-        updateActivity.setEnd_time(activityDetail.getEnd_time());
+        updateAndDeleteActivityStatus();
+        activity.setName(activityDetail.getName());
+        activity.setDescription(activityDetail.getDescription());
+        activity.setStart_time(activityDetail.getStart_time());
+        activity.setEnd_time(activityDetail.getEnd_time());
 
-        activityRepository.save(updateActivity);
+        activityRepository.save(activity);
         return SimpleResponseDTO.builder().success(true).message("Activity edited").build();
     }
 
@@ -111,17 +136,21 @@ public class ActivityServiceImpl implements ActivityService {
         Activity activity = activityRepository.findFirstById(activityId);
         Group group = activity.getGroup();
         long groupId = group.getId();
+        // Validate user and group
         SimpleResponseDTO checkResult = checkUserAndGroup(groupId);
         if (!checkResult.isSuccess()) {
             return checkResult;
         }
+        // Validate activity
+        SimpleResponseDTO validateResult = validateActivity(activity, groupId);
+        if (!validateResult.isSuccess()) {
+            return validateResult;
+        }
         for (User user : activity.getUsers()) {
             user.getActivities().remove(activity);
         }
-        updateActivityStatus();
         group.getActivities().remove(activity);
         activityRepository.delete(activity);
         return SimpleResponseDTO.builder().success(true).message("Activity deleted").build();
     }
-
 }
